0 bugs editar_usuario

+++ pages/editar_usuario.php
@@ -21,19 +21,22 @@ $erro = false;
 $dados_editar = $objUsuario->getDadosById($_SESSION['id_user'])[0];  
 $foto = $dados_editar['foto'];
 
-if(isset($_FILES['foto'])){ 
-    
-    
-    
-    $name_img = $_FILES['foto']['name'];
-    $new_name  = uniqid(). '-' . substr($name_img, 0, 20);
-    $path = '../assets/imgs/users/';
-    move_uploaded_file($_FILES['foto']['tmp_name'], $path.$new_name);
-    
-    unlink($path . $dados_editar['foto']);
+if(isset($_FILES['foto'])){
 
-    $objUsuario->setImage($dados_editar['email'],$new_name);
-  
+    if(strlen($_FILES['foto']['name']) > 0 ){ 
+      
+         
+        $name_img = $_FILES['foto']['name'];
+        $new_name  = uniqid(). '-' . substr($name_img, 0, 20);
+        $path = '../assets/imgs/users/';
+        move_uploaded_file($_FILES['foto']['tmp_name'], $path.$new_name);
+        
+        unlink($path . $dados_editar['foto']);
+    
+        $objUsuario->setImage($dados_editar['email'],$new_name);
+      
+
+    }
 }
 if(isset($_POST['btn_submit'])){
